+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Current
 
+- We have now `tradingstrategey.client.BaseClient` as the base class for different client implementations
 - Adding `UniswapV2MockClient` implementation that mimics `Client`, but reads all data
   from an EVM test backend. Makes live trading tests much easier.
-- 
+- Small improvements to PairUniverse
+
 # 0.13.10
 
 - Fix detached indicator charting bugs for different `volume_bar_modes`